osinstancectl: Improve compatibility with openslides versions
Currently, 'openslides' creates a db-data directory which we don't need
and have to remove.  Future versions will avoid creating the useless
directory, at which point we can remove the code.  For now, however,
this patch makes osinstancectl compatible with both old and new versions
of 'openslides'.

+++ osinstancectl.sh
@@ -486,7 +486,10 @@ create_instance_dir() {
 
   # Due to a bug in "openslides", the db-data directory is created even if the
   # stack's Postgres service that would require it is disabled.
-  if [[ $(value_from_config_yml "$PROJECT_DIR" '.disablePostgres') == "true" ]]; then
+  # XXX: This is going to be fixed in the near future.  For now, remain
+  # backwards-compatible.
+  if [[ -d "${PROJECT_DIR}/db-data" ]] && [[ $(value_from_config_yml "$PROJECT_DIR" '.disablePostgres') == "true" ]]
+  then
     rmdir "${PROJECT_DIR}/db-data"
   fi
